ugc/pyproject.toml
[tool.poetry]
name = "ugc"
version = "0.1.0"
description = "UGC analytics service"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = {extras = ["async"], version = "^3.0.2"}
pydantic-settings = "^2.2.1"
flask-jwt-extended = "^4.6.0"
pydantic = "^2.6.3"
kafka-python = "^2.0.2"
gunicorn = "^21.2.0"
pytest = "^8.1.1"
requests = "^2.31.0"
flask-pymongo = "^2.3.0"
sentry-sdk = "^1.44.0"
motor = "^3.4.0"
aiokafka = "^0.10.0"
pytest-asyncio = "^0.23.6"


[tool.poetry.group.dev.dependencies]
ipython = "^8.22.2"
blue = "^0.9.1"
poetry-plugin-export = "^1.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
